refactor: modularize navigation and enhance hero
