+++ lib/main.dart
@@ -1,21 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'theme/app_theme.dart';
-import 'package:tour_flutter_main/models/auth_models.dart';
-import 'screens/auth/login_screen.dart';
-import 'screens/auth/register_screen.dart';
-import 'screens/tours/tour_list_screen.dart';
-import 'screens/home/home_web_screen.dart';
-import 'screens/cars/car_list_screen.dart';
-import 'screens/admin/admin_tour_create_screen.dart';
-import 'screens/admin/admin_panel_screen.dart';
-import 'screens/profile/profile_screen.dart';
-import 'screens/booking/booking_screen.dart';
-import 'screens/recommendation/recommendation_screen.dart';
-import 'services/auth_service.dart';
 import 'services/stripe_service.dart';
-import 'widgets/resizable_navigation_rail.dart';
 import 'widgets/app_layout.dart';
+import 'navigation/auth_wrapper.dart';
+import 'routes/app_routes.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -42,320 +31,7 @@ class MyApp extends StatelessWidget {
       themeMode: ThemeMode.system,
       builder: (context, child) => AppLayout(child: child ?? const SizedBox()),
       home: const AuthWrapper(),
-      routes: {
-        '/login': (context) => const LoginScreen(),
-        '/register': (context) => const RegisterScreen(),
-        '/home': (context) => const HomeWebScreen(),
-        '/tours': (context) => const TourListScreen(),
-        '/cars': (context) => const CarListScreen(),
-        '/admin-panel': (context) => const AdminPanelScreen(),
-        '/admin/create-tour': (context) => const AdminTourCreateScreen(),
-        '/profile': (context) => const ProfileScreen(),
-        '/bookings': (context) => const BookingScreen(),
-        '/recommendations': (context) => const RecommendationScreen(),
-      },
-    );
-  }
-}
-
-class AuthWrapper extends StatefulWidget {
-  const AuthWrapper({super.key});
-
-  @override
-  State<AuthWrapper> createState() => _AuthWrapperState();
-}
-
-class _AuthWrapperState extends State<AuthWrapper>
-    with TickerProviderStateMixin {
-  final AuthService _authService = AuthService();
-  bool _isLoading = true;
-  bool _isLoggedIn = false;
-  late AnimationController _loadingController;
-  late Animation<double> _pulseAnimation;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadingController = AnimationController(
-      duration: const Duration(seconds: 2),
-      vsync: this,
-    );
-    _pulseAnimation = Tween<double>(begin: 0.8, end: 1.2).animate(
-      CurvedAnimation(parent: _loadingController, curve: Curves.easeInOut),
-    );
-    _loadingController.repeat(reverse: true);
-    _checkAuthStatus();
-  }
-
-  @override
-  void dispose() {
-    _loadingController.dispose();
-    super.dispose();
-  }
-
-  Future<void> _checkAuthStatus() async {
-    try {
-      final isLoggedIn = await _authService.isLoggedIn();
-      if (mounted) {
-        setState(() {
-          _isLoggedIn = isLoggedIn;
-          _isLoading = false;
-        });
-        _loadingController.stop();
-      }
-    } catch (e) {
-      if (mounted) {
-        setState(() {
-          _isLoggedIn = false;
-          _isLoading = false;
-        });
-        _loadingController.stop();
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (_isLoading) {
-      return Scaffold(
-        body: Container(
-          decoration: BoxDecoration(
-            gradient: LinearGradient(
-              begin: Alignment.topCenter,
-              end: Alignment.bottomCenter,
-              colors: [
-                Theme.of(context).colorScheme.primary.withOpacity(0.1),
-                Theme.of(context).colorScheme.background,
-              ],
-            ),
-          ),
-          child: Center(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                AnimatedBuilder(
-                  animation: _pulseAnimation,
-                  builder: (context, child) {
-                    return Transform.scale(
-                      scale: _pulseAnimation.value,
-                      child: Container(
-                        padding: const EdgeInsets.all(32),
-                        decoration: BoxDecoration(
-                          color: Theme.of(context).colorScheme.surface,
-                          shape: BoxShape.circle,
-                          boxShadow: [
-                            BoxShadow(
-                              color: Theme.of(
-                                context,
-                              ).colorScheme.primary.withOpacity(0.2),
-                              blurRadius: 20,
-                              spreadRadius: 5,
-                            ),
-                          ],
-                        ),
-                        child: Icon(
-                          Icons.travel_explore_rounded,
-                          size: 64,
-                          color: Theme.of(context).colorScheme.primary,
-                        ),
-                      ),
-                    );
-                  },
-                ),
-                const SizedBox(height: 32),
-                Text(
-                  'TourApp',
-                  style: Theme.of(context).textTheme.headlineLarge?.copyWith(
-                    color: Theme.of(context).colorScheme.primary,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-                const SizedBox(height: 8),
-                Text(
-                  'Your next adventure awaits...',
-                  style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                    color: Theme.of(
-                      context,
-                    ).colorScheme.onSurface.withOpacity(0.7),
-                  ),
-                ),
-                const SizedBox(height: 32),
-                SizedBox(
-                  width: 32,
-                  height: 32,
-                  child: CircularProgressIndicator(
-                    strokeWidth: 3,
-                    color: Theme.of(context).colorScheme.primary,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      );
-    }
-
-    return _isLoggedIn ? const HomeScreen() : const LoginScreen();
-  }
-}
-
-class HomeScreen extends StatefulWidget {
-  const HomeScreen({super.key});
-
-  @override
-  State<HomeScreen> createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
-  final AuthService _authService = AuthService();
-  int _currentIndex = 0;
-  bool _isAdmin = false;
-
-  final List<Widget> _userScreens = [
-    const HomeWebScreen(),
-    const TourListScreen(),
-    const RecommendationScreen(),
-    const CarListScreen(),
-    const BookingScreen(),
-    const ProfileScreen(),
-  ];
-
-  final List<Widget> _adminScreens = [
-    const HomeWebScreen(),
-    const TourListScreen(),
-    const RecommendationScreen(),
-    const CarListScreen(),
-    const BookingScreen(),
-    const AdminPanelScreen(),
-    const ProfileScreen(),
-  ];
-
-  @override
-  void initState() {
-    super.initState();
-    _checkAdminStatus();
-  }
-
-  Future<void> _checkAdminStatus() async {
-    final isAdmin = await _authService.isAdmin();
-    if (mounted) {
-      setState(() {
-        _isAdmin = isAdmin;
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final screens = _isAdmin ? _adminScreens : _userScreens;
-    final colorScheme = Theme.of(context).colorScheme;
-    final bool isDesktop = MediaQuery.of(context).size.width >= 800;
-
-    final navigationDestinations = <NavigationDestination>[
-      NavigationDestination(
-        icon: const Icon(Icons.home_outlined),
-        selectedIcon: Icon(Icons.home, color: colorScheme.primary),
-        label: 'Home',
-      ),
-      NavigationDestination(
-        icon: const Icon(Icons.explore_outlined),
-        selectedIcon: Icon(Icons.explore, color: colorScheme.primary),
-        label: 'Discover',
-      ),
-      NavigationDestination(
-        icon: const Icon(Icons.recommend_outlined),
-        selectedIcon: Icon(Icons.recommend, color: colorScheme.primary),
-        label: 'For You',
-      ),
-      NavigationDestination(
-        icon: const Icon(Icons.directions_car_outlined),
-        selectedIcon: Icon(Icons.directions_car, color: colorScheme.primary),
-        label: 'Cars',
-      ),
-      NavigationDestination(
-        icon: const Icon(Icons.bookmark_border_outlined),
-        selectedIcon: Icon(Icons.bookmark, color: colorScheme.primary),
-        label: 'Bookings',
-      ),
-      if (_isAdmin)
-        NavigationDestination(
-          icon: const Icon(Icons.admin_panel_settings_outlined),
-          selectedIcon: Icon(
-            Icons.admin_panel_settings,
-            color: colorScheme.primary,
-          ),
-          label: 'Admin',
-        ),
-      NavigationDestination(
-        icon: const Icon(Icons.person_outline),
-        selectedIcon: Icon(Icons.person, color: colorScheme.primary),
-        label: 'Profile',
-      ),
-    ];
-
-    return Scaffold(
-      body: isDesktop
-          ? Row(
-              children: [
-                ResizableNavigationRail(
-                  selectedIndex: _currentIndex,
-                  onDestinationSelected: (index) {
-                    if (index != _currentIndex) {
-                      setState(() => _currentIndex = index);
-                      HapticFeedback.selectionClick();
-                    }
-                  },
-                  backgroundColor: colorScheme.surface,
-                  selectedIconTheme: IconThemeData(color: colorScheme.primary),
-                  indicatorColor: colorScheme.primary.withOpacity(0.1),
-                  destinations: navigationDestinations
-                      .map(
-                        (e) => NavigationRailDestination(
-                          icon: e.icon,
-                          selectedIcon: e.selectedIcon!,
-                          label: Text(e.label),
-                        ),
-                      )
-                      .toList(),
-                ),
-                const VerticalDivider(width: 1),
-                Expanded(
-                  child: IndexedStack(index: _currentIndex, children: screens),
-                ),
-              ],
-            )
-          : IndexedStack(index: _currentIndex, children: screens),
-      bottomNavigationBar: isDesktop
-          ? null
-          : Container(
-              decoration: BoxDecoration(
-                color: colorScheme.surface,
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.black.withOpacity(0.1),
-                    blurRadius: 8,
-                    offset: const Offset(0, -2),
-                  ),
-                ],
-              ),
-              child: SafeArea(
-                child: NavigationBar(
-                  selectedIndex: _currentIndex,
-                  onDestinationSelected: (index) {
-                    if (index != _currentIndex) {
-                      setState(() {
-                        _currentIndex = index;
-                      });
-                      HapticFeedback.selectionClick();
-                    }
-                  },
-                  backgroundColor: colorScheme.surface,
-                  surfaceTintColor: colorScheme.surface,
-                  indicatorColor: colorScheme.primary.withOpacity(0.1),
-                  destinations: navigationDestinations,
-                ),
-              ),
-            ),
+      routes: AppRoutes.routes,
     );
   }
 }
