+++ lib/screens/home/widgets/hero_section.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../../../widgets/glass_container.dart';
 
 class HomeHeroSection extends StatelessWidget {
   const HomeHeroSection({super.key});
@@ -34,39 +35,45 @@ class HomeHeroSection extends StatelessWidget {
                 ),
               ),
               Center(
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Text(
-                      'Explore the World',
-                      style: theme.textTheme.displaySmall?.copyWith(
-                        color: Colors.white,
-                        fontWeight: FontWeight.bold,
-                      ),
-                      textAlign: TextAlign.center,
-                    ),
-                    const SizedBox(height: 16),
-                    Text(
-                      'Find unforgettable experiences and make memories.',
-                      style: theme.textTheme.titleMedium?.copyWith(
-                        color: Colors.white70,
-                      ),
-                      textAlign: TextAlign.center,
-                    ),
-                    const SizedBox(height: 32),
-                    FilledButton(
-                      onPressed: () {
-                        Navigator.of(context).pushNamed('/tours');
-                      },
-                      style: FilledButton.styleFrom(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 32,
-                          vertical: 20,
+                child: GlassContainer(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 24,
+                    vertical: 32,
+                  ),
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      Text(
+                        'Explore the World',
+                        style: theme.textTheme.displaySmall?.copyWith(
+                          color: Colors.white,
+                          fontWeight: FontWeight.bold,
                         ),
+                        textAlign: TextAlign.center,
                       ),
-                      child: const Text('Start Exploring'),
-                    ),
-                  ],
+                      const SizedBox(height: 16),
+                      Text(
+                        'Find unforgettable experiences and make memories.',
+                        style: theme.textTheme.titleMedium?.copyWith(
+                          color: Colors.white70,
+                        ),
+                        textAlign: TextAlign.center,
+                      ),
+                      const SizedBox(height: 32),
+                      FilledButton(
+                        onPressed: () {
+                          Navigator.of(context).pushNamed('/tours');
+                        },
+                        style: FilledButton.styleFrom(
+                          padding: const EdgeInsets.symmetric(
+                            horizontal: 32,
+                            vertical: 20,
+                          ),
+                        ),
+                        child: const Text('Start Exploring'),
+                      ),
+                    ],
+                  ),
                 ),
               ),
             ],
