+++ lib/widgets/glass_container.dart
@@ -1,0 +1,39 @@
+import 'dart:ui';
+import 'package:flutter/material.dart';
+
+/// A reusable frosted glass container that can wrap any widget.
+class GlassContainer extends StatelessWidget {
+  final Widget child;
+  final EdgeInsets padding;
+  final double blur;
+  final double opacity;
+  final BorderRadius borderRadius;
+
+  const GlassContainer({
+    super.key,
+    required this.child,
+    this.padding = const EdgeInsets.all(16),
+    this.blur = 10,
+    this.opacity = 0.2,
+    this.borderRadius = const BorderRadius.all(Radius.circular(16)),
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return ClipRRect(
+      borderRadius: borderRadius,
+      child: BackdropFilter(
+        filter: ImageFilter.blur(sigmaX: blur, sigmaY: blur),
+        child: Container(
+          padding: padding,
+          decoration: BoxDecoration(
+            color: Colors.white.withOpacity(opacity),
+            borderRadius: borderRadius,
+            border: Border.all(color: Colors.white.withOpacity(0.3)),
+          ),
+          child: child,
+        ),
+      ),
+    );
+  }
+}
